server/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# MQTT client
paho-mqtt==1.6.1

# MongoDB driver
motor==3.3.2
pymongo==4.6.0

# AI/ML for forecasting
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Alerts
requests==2.31.0
python-telegram-bot==20.7

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
